inc: work on pr 2.2
