+++ students/k33401/Reingeverts_Vadim/.djlintrc
@@ -1,0 +1,3 @@
+{
+    "ignore": "H030,H031"
+}
