+++ students/k33401/Reingeverts_Vadim/Lr2/Pr/django_project_reingeverts/project_first_app/admin.py
@@ -4,6 +4,6 @@ from django.contrib import admin
 from .models import Car, Ownership, CarOwner, DriverLicense
 
 admin.site.register(Car)
-admin.site.register(Ownership)
 admin.site.register(CarOwner)
+admin.site.register(Ownership)
 admin.site.register(DriverLicense)
