+++ students/k33401/Reingeverts_Vadim/Lr2/Pr/django_project_reingeverts/templates/car_owner_detail.html
@@ -1,0 +1,9 @@
+{% extends "./base.html" %}
+{% block title %}
+    {{ title }}
+{% endblock title %}
+{% block main %}
+    <h1>Car owner details</h1>
+    <p>Name: {{ car_owner.first_name }} {{ car_owner.last_name }}</p>
+    <p>Date of birth: {{ car_owner.date_of_birth|date:"Y.m.d" }}</p>
+{% endblock main %}
