--- students/k33401/Reingeverts_Vadim/Lr2/Pr1-5/django_project_reingeverts/project_first_app/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path('owner/<int:car_owner_pk>',
-         views.car_owner_detail, name='car-owner-detail'),
-]
