+++ students/k33401/Reingeverts_Vadim/Lr2/Readme.md
@@ -1,10 +1,6 @@
-## Practical work 1-5
+# Lab work 2. Implementation of a simple website on django.
 
-Model
-![](https://i.imgur.com/60P88U7.png)
-
-View
-![](https://i.imgur.com/G0y3Vvm.png)
+## General
 
 ### Activating python venv
 
@@ -12,14 +8,26 @@ View
 source ../.web-dev-env/Scripts/activate
 ```
 
-> VSCode:
->
-> `Python: Select Interpreter` – to fix IntelliSense
+### VSCode setup
+
+`Python: Select Interpreter` – to fix IntelliSense
+
+```bash
+python -m pip install --upgrade djlint
+```
+
+Install djlint extension and add its config - `.djlintrc`:
+
+```json
+{
+    "ignore": "H030,H031"
+}
+```
 
 ### Running
 
 ```bash
-python django_project_reingeverts/manage.py runserver
+python django_project_reingeverts/manage.py makemigrations && python django_project_reingeverts/manage.py migrat e && python django_project_reingeverts/manage.py runserver
 ```
 
 ### Admin
@@ -29,3 +37,19 @@ User: `admin`
 Email: `admin@example.com`
 
 Password: `admin`
+
+## Practical part
+
+### Practical work 2.1.1-2.1.5
+
+Model
+![](https://i.imgur.com/60P88U7.png)
+
+View
+![](https://i.imgur.com/G0y3Vvm.png)
+
+### Practical work 2.2
+
+## Lab work part
+
+...
